Add graceful shutdown to TcpServer
This commit prototypes support for graceful shutdown in `TcpServer`
through some new methods. A `TcpServer::bind` method was acced which is
used to construct a server and unlike `serve` returns the value
immediately.

The returned value can then be used to inspect the state of the server,
finally calling `run` to execute indefinitely or `run_until` to run
until a shutdown signal is received.